organize lib
